+++ src/com/journaldev/collections/PriorityQueueExample.java
@@ -30,7 +30,7 @@ public class PriorityQueueExample {
     public static Comparator<Customer> idComparator = new Comparator<Customer>() {
         @Override
         public int compare(Customer c1, Customer c2) {
-            return (int) (c1.getId() - c2.getId());
+            return (c1.getId() - c2.getId());
         }
     };
